refactor: restructure project for better documentation
