+++ README.md
@@ -1,7 +1,7 @@
 A library for interacting with SeisComP, the modular system for processing and distributing seismic data. Runs on Node
 and in the browser. Developed with [bun.js](https://github.com/oven-sh/bun) (a Node.js alternative using
 JavaScriptCore - you should try it! it's faster than Node and you don't have to build your TypeScript project to test
-it).
+it). Also fully typed.
 
 ## Installation
 
@@ -29,11 +29,16 @@ an issue.
 ```js
 import { FDSNWS } from 'seiscomp'
 
-const client = new FDSNWS('https://moho.iag.usp.br/fdsnws')
 
-const events = await client.station.queryStation({
-  networkCode: 'IU',
-  channelCode: 'BHZ'
-})
+const client = new FDSNWS('https://moho.iag.usp.br/fdsnws/')
+
+await client.station.queryNetwork('BL')
+
+await client.station.query()
+  .channel.network('BR')
+  .channel.station('VIL?')
+  .time.startAfter('2013-01-01')
+  .finish()
+  .then(console.log)
 ```
