+++ examples/fdsnws.js
@@ -1,4 +1,4 @@
-import { FDSNWS } from '../src/index.ts' // this will not work if you are on node.js!
+import { FDSNWS } from '../src/index.js' // this will not work if you are on node.js!
 
 const client = new FDSNWS('https://moho.iag.usp.br/fdsnws/')
