+++ src/modules/fdsnws.ts
@@ -1,22 +1,28 @@
-import { Network, Station, StationQuery } from '../../types/fdsnws.js'
-import { sc3mlQueryToNetwork, sc3mlStationToStation, stationQueryToQuery, toArray } from '../../utilities.js'
-import FDSNWS from './index.js'
-import FDSNWSService from '../../structures/FDSNWSService.js'
-import { QueryBuilder } from '../../builders/queryBuilder.js'
+import SeisModule from '../structures/SeisModule.js'
+import { get } from '../http.js'
+import FDSNWSService from '../structures/FDSNWSService.js'
+import { QueryBuilder } from '../builders/queryBuilder.js'
+import { Network, Station, StationQuery } from '../types/fdsnws.js'
+import { deprecate } from 'util'
+import { sc3mlQueryToNetwork, sc3mlStationToStation, stationQueryToQuery, toArray } from '../utilities.js'
 
 /**
  * The StationService class, used to interact with the `fdsnws_station` service.
  *
  * ⚠ The `station` service may **not** be available on your desired SeisComP instance.
  *
+ * ⚠ Please note that using features (such as some constraints) that aren't enabled on the server **will** result in an error.
+ *
  * If you are experiencing issues, please check if the service is available on the desired instance by visiting the FDSNWS URL in your browser.
  * Contact the system administrator for further assistance.
  *
  * @see https://www.seiscomp.de/doc/apps/fdsnws.html#station
  */
-export class StationService extends FDSNWSService {
+class StationService extends FDSNWSService {
+  /** @hidden */
   private fdsnws: FDSNWS
 
+  /** @hidden */
   constructor (fdsnws: FDSNWS) {
     super()
     this.fdsnws = fdsnws
@@ -25,10 +31,10 @@ export class StationService extends FDSNWSService {
   /**
    * Returns a query builder for the `station` service.
    * Refer to the query builder documentation for usage.
+   * {@link builders/queryBuilder.QueryBuilder QueryBuilder documentation}
    *
-   * ⚠ Using features that aren't enabled on the server **will** result in an error.
+   * ⚠ Please note that using features that aren't enabled on the server **will** result in an error.
    *
-   * {@link builders.queryBuilder.QueryBuilder}
    */
   query (): QueryBuilder<Station[]> {
     return new QueryBuilder(
@@ -37,9 +43,13 @@ export class StationService extends FDSNWSService {
     )
   }
 
-
-  /** Queries stations from the FDSNWS service. */
+  /**
+   * Queries stations from the FDSNWS service
+   * @deprecated Use {@link StationService.query} instead. This method will be removed in a future release.
+   */
   async queryStation (data: StationQuery): Promise<Station[]> {
+    deprecate(() => {
+    }, 'queryStation is deprecated. Use query instead. This method will be removed in a future release.')
     return this._queryStation(stationQueryToQuery(data))
   }
 
@@ -52,7 +62,6 @@ export class StationService extends FDSNWSService {
       })
       .then((r) => r ? toArray(r.seiscomp.Inventory.network.station).map(sc3mlStationToStation) : [])
   }
-
 
   /**
    * Returns data about a network.
@@ -84,3 +93,36 @@ export class StationService extends FDSNWSService {
     return this.fdsnws._request<string>('/station/1/version', { format: 'text' })
   }
 }
+
+/**
+ * The FDSNWS module, used to interact with the FDSNWS services.
+ * @see https://www.seiscomp.de/doc/apps/fdsnws.html
+ *
+ * ⚠ Please note that using features that aren't enabled on the server **will** result in an error.
+ * ⚠ The FDSNWS server may **not** be available on your desired SeisComP instance.
+ */
+class FDSNWS extends SeisModule {
+  /** The URL of the FDSNWS server. */
+  url: string
+
+  /** The StationService class, used to interact with the fdsnws_station service. */
+  station = new StationService(this)
+
+  /**
+   * Creates a new FDSNWS module.
+   */
+  public constructor (url: string) {
+    super()
+    this.url = url.endsWith('/') ? url.slice(0, -1) : url
+  }
+
+  /** @hidden */
+  _request<T> (path: string, query: Record<string, any> | string, xml: boolean = false): Promise<T | undefined> {
+    return get(this.url + path, { query, xml }).catch((err) => {
+      if (err.status === 404) return undefined
+      throw err
+    }) as Promise<T | undefined>
+  }
+}
+
+export { FDSNWS, StationService }
